*: move contracts into a dir of their own
It's a bit hard to navigate here now. "common" stays as is to minimize import
changes.
