+++ contracts/alphabet/alphabet_contract.go
