+++ contracts/alphabet/config.yml
