+++ contracts/neofs/config.yml
