+++ contracts/neofs/doc.go
