+++ contracts/neofs/neofs_contract.go
