+++ contracts/neofsid/neofsid_contract.go
