+++ contracts/netmap/netmap_contract.go
