+++ contracts/nns/migration_test.go
