+++ contracts/nns/nns_contract.go
