+++ contracts/nns/recordtype.go
